+++ build_cli/test/src/pubviz_example.dart
@@ -26,7 +26,7 @@ class PeanutOptions {
 enum FormatOptions { dot, html }
 
 const _formatOptionsHelp = const <FormatOptions, String>{
-  FormatOptions.dot: 'Generate a GraphViz dot file',
+  FormatOptions.dot: "Generate a GraphViz 'dot' file.",
   FormatOptions.html:
       'Wrap the GraphViz dot format in an HTML template which renders it.'
 };